refactor(slice-25b): unify MongoDB entry points — start-services.sh mongodb subcommands + mongodb-setup.md
Consolidates local-atlas.sh and dual setup docs into one script and one guide;
compose overlay merged into docker-compose.yml env vars. PROGRESS and CHANGELOG updated.
