+++ .env.example
@@ -3,7 +3,7 @@
 # Voyage sweep (example-mongodb-voyage.yaml): MONGODB_URI + VOYAGE_API_KEY + Tier 1 limits below
 # SIE sweep (example-mongodb-sie.yaml):       MONGODB_URI + SIE_ENABLED=true + SIE_ENDPOINT
 #                                             (+ SIE_API_KEY when gateway auth is enabled)
-# Full checklist: docs/user-guide/cloud-setup.md#before-you-run-a-sweep
+# Full checklist: docs/user-guide/mongodb-setup.md#before-you-run-a-sweep
 VOYAGE_API_KEY=your_voyage_api_key_here
 # Voyage rate limits — free tier defaults below; override for Tier 1 (Voyage sweep)
 VOYAGE_RPM_LIMIT=3
@@ -19,10 +19,10 @@ VOYAGE_TPM_LIMIT=10000
 MONGODB_URI=your_mongodb_atlas_uri_here
 #
 # Option B — MongoDB Atlas Local (local Docker, no cloud account needed):
-#   Start: docker compose -f docker-compose.yml -f docker-compose.local-atlas.yml --profile local-atlas up -d
+#   Start: ./start-services.sh --local
 #   Then set this URI for host CLI and local server:
 # MONGODB_URI=mongodb://localhost:27017/rag_params_finder?directConnection=true
-#   See docs/user-guide/local-atlas-setup.md
+#   See docs/user-guide/mongodb-setup.md
 
 # Optional — auto-detect cluster storage quota for the dashboard (Atlas Admin API).
 # Without these, set MONGODB_STORAGE_LIMIT_MB manually or quota UI stays hidden.
