+++ docs/README.md
@@ -14,12 +14,12 @@ All guides for **rag-params-finder**, organized by **who you are** and **what yo
 
 | Persona | Start here | Then |
 |---------|------------|------|
-| **New user — cloud accounts** | [user-guide/cloud-setup.md](./user-guide/cloud-setup.md) | [QUICKSTART.md](../QUICKSTART.md) → [user-guide/getting-started.md](./user-guide/getting-started.md) |
+| **New user — MongoDB + providers** | [user-guide/mongodb-setup.md](./user-guide/mongodb-setup.md) | [QUICKSTART.md](../QUICKSTART.md) → [user-guide/getting-started.md](./user-guide/getting-started.md) |
 | **New user — first sweep** | [QUICKSTART.md](../QUICKSTART.md) | [user-guide/getting-started.md](./user-guide/getting-started.md) → dashboard at `http://localhost:5374` |
 | **Operator — config & CLI** | [user-guide/configuration.md](./user-guide/configuration.md) | [user-guide/cli-reference.md](./user-guide/cli-reference.md) |
 | **Operator — dashboard** | [user-guide/dashboard-guide.md](./user-guide/dashboard-guide.md) | [user-guide/configuration.md](./user-guide/configuration.md) (tiebreaker, env vars) |
 | **Operator — SIE (BGE-M3 / Stella / SPLADE)** | [user-guide/sie-setup.md](./user-guide/sie-setup.md) | [user-guide/troubleshooting.md](./user-guide/troubleshooting.md#sie-superlinked-inference-engine) |
-| **Operator — fixing errors** | [user-guide/troubleshooting.md](./user-guide/troubleshooting.md) | [user-guide/cloud-setup.md](./user-guide/cloud-setup.md) (indexes, Voyage tiers) |
+| **Operator — fixing errors** | [user-guide/troubleshooting.md](./user-guide/troubleshooting.md) | [user-guide/mongodb-setup.md](./user-guide/mongodb-setup.md) (indexes, Voyage tiers) |
 | **Contributor — system design** | [contributor-guide/architecture.md](./contributor-guide/architecture.md) | [adr/](./adr/) |
 | **Contributor — extending** | [contributor-guide/extending.md](./contributor-guide/extending.md) | [contributor-guide/development.md](./contributor-guide/development.md) |
 | **Contributor — dev environment** | [contributor-guide/development.md](./contributor-guide/development.md) | [slices/PROGRESS.md](./slices/PROGRESS.md) · [slices/](./slices/) specs |
@@ -31,7 +31,7 @@ All guides for **rag-params-finder**, organized by **who you are** and **what yo
 
 | Doc | What it covers |
 |-----|----------------|
-| [user-guide/cloud-setup.md](./user-guide/cloud-setup.md) | MongoDB Atlas + optional Voyage AI accounts, search indexes |
+| [user-guide/mongodb-setup.md](./user-guide/mongodb-setup.md) | MongoDB Atlas cloud or local Docker, Voyage AI, search indexes |
 | [user-guide/getting-started.md](./user-guide/getting-started.md) | Install, configure, first experiment (step-by-step) |
 | [user-guide/sie-setup.md](./user-guide/sie-setup.md) | SIE setup — remote gateway (preferred) or optional self-hosted Docker; warm-up, Aim UI, known issues |
 | [user-guide/configuration.md](./user-guide/configuration.md) | Full YAML config reference, env vars, sweep dimensions |
@@ -83,7 +83,7 @@ All guides for **rag-params-finder**, organized by **who you are** and **what yo
 | Task | Doc / command |
 |------|----------------|
 | Install and run first sweep | [QUICKSTART.md](../QUICKSTART.md) |
-| Atlas vector + text search indexes | [user-guide/cloud-setup.md](./user-guide/cloud-setup.md) |
+| Atlas vector + text search indexes | [user-guide/mongodb-setup.md](./user-guide/mongodb-setup.md) |
 | Example YAML configs | `configs/example-mongodb-local.yaml`, `configs/example-mongodb-voyage.yaml`, `configs/example-mongodb-sie.yaml` |
 | Quality gates before commit | [contributor-guide/development.md](./contributor-guide/development.md) · `./scripts/quality-gates.sh` |
 | Docker server + dashboard | [slices/SLICE-14-DOCKER-COMPOSE.md](./slices/SLICE-14-DOCKER-COMPOSE.md) |
