+++ start-services.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 # Start rag-params-finder server + dashboard via Docker Compose (prod stack)
-# Usage: ./start-services.sh [--force-build] [--local]
-#   Rebuilds images only when build context changed, images are missing, or --force-build.
-#   Env: RAG_FORCE_BUILD=1 (same as --force-build), RAG_DEV_STACK=1, RAG_LOCAL_ATLAS=1, NONINTERACTIVE=1
+# Usage:
+#   ./start-services.sh [--force-build] [--local]     Start full stack
+#   ./start-services.sh mongodb start|stop|reset|status   MongoDB Atlas Local container only
+#   Env: RAG_FORCE_BUILD=1, RAG_DEV_STACK=1, RAG_LOCAL_ATLAS=1, NONINTERACTIVE=1
 set -e
 set -o pipefail
 
@@ -13,22 +14,29 @@ cd "$SCRIPT_DIR"
 source ./scripts/docker-cleanup.sh
 # shellcheck source=scripts/docker-build-context.sh
 source ./scripts/docker-build-context.sh
+# shellcheck source=scripts/lib/compose.sh
+source ./scripts/lib/compose.sh
 
 FORCE_BUILD=0
 LOCAL_ATLAS=0
 
-LOCAL_URI="mongodb://localhost:27017/rag_params_finder?directConnection=true"
-
 usage() {
   cat <<EOF
 Usage: ./start-services.sh [OPTIONS]
+       ./start-services.sh mongodb start|stop|reset|status
 
-Start server + dashboard via Docker Compose.
+Start server + dashboard via Docker Compose (default), or manage MongoDB Atlas Local only.
 
-Options:
+Stack options:
   --local, -l                  Use MongoDB Atlas Local (no cloud account needed)
   --force-build, --build, -b   Rebuild images even when build context is unchanged
   -h, --help                   Show this help
+
+MongoDB container only (Atlas Local Docker):
+  mongodb start                Start mongodb-local container
+  mongodb stop                 Stop container
+  mongodb reset                Stop + wipe data volume
+  mongodb status               Container state + CLI connection string
 
 Environment:
   RAG_LOCAL_ATLAS=1            Same as --local
@@ -39,9 +47,66 @@ Environment:
 Backends:
   Cloud (default):  requires MONGODB_URI in .env (mongodb+srv://...)
   Local (--local):  starts mongodb-atlas-local container; no .env MONGODB_URI needed
-                    CLI: export MONGODB_URI="$LOCAL_URI"
-                    Manage container standalone: ./scripts/local-atlas.sh
+                    CLI: export MONGODB_URI="$RAG_LOCAL_MONGODB_URI_HOST"
 EOF
+}
+
+cmd_mongodb_start() {
+  echo "Starting MongoDB Atlas Local..."
+  "${DOCKER_COMPOSE[@]}" "${COMPOSE_FILES[@]}" "${COMPOSE_PROFILES[@]}" up -d mongodb-local
+  echo ""
+  echo "Waiting for MongoDB Atlas Local to be ready..."
+  wait_for_mongodb_local_healthy
+  print_local_atlas_cli_hints 1
+}
+
+cmd_mongodb_stop() {
+  echo "Stopping MongoDB Atlas Local..."
+  "${DOCKER_COMPOSE[@]}" "${COMPOSE_FILES[@]}" "${COMPOSE_PROFILES[@]}" stop mongodb-local
+  echo "Stopped."
+}
+
+cmd_mongodb_reset() {
+  echo "Stopping and wiping MongoDB Atlas Local data volume..."
+  "${DOCKER_COMPOSE[@]}" "${COMPOSE_FILES[@]}" "${COMPOSE_PROFILES[@]}" rm -sf mongodb-local
+  docker volume rm "$RAG_MONGODB_LOCAL_VOLUME" 2>/dev/null || true
+  echo "Volume wiped. Run './start-services.sh mongodb start' to recreate."
+}
+
+cmd_mongodb_status() {
+  local state
+  state="$(docker inspect --format='{{.State.Status}}' "$RAG_MONGODB_LOCAL_CONTAINER" 2>/dev/null || echo "not found")"
+  local health
+  health="$(docker inspect --format='{{.State.Health.Status}}' "$RAG_MONGODB_LOCAL_CONTAINER" 2>/dev/null || echo "—")"
+  echo "Container: $RAG_MONGODB_LOCAL_CONTAINER"
+  echo "  State:  $state"
+  echo "  Health: $health"
+  if [[ "$state" == "running" && "$health" == "healthy" ]]; then
+    print_local_atlas_cli_hints 1
+  fi
+}
+
+run_mongodb_subcommand() {
+  if ! command -v docker >/dev/null 2>&1; then
+    echo "Docker is not installed. See https://docs.docker.com/get-docker/" >&2
+    exit 1
+  fi
+  compose_detect
+  compose_files
+  compose_local_atlas_profiles
+
+  local cmd="${1:-start}"
+  case "$cmd" in
+    start)  cmd_mongodb_start ;;
+    stop)   cmd_mongodb_stop ;;
+    reset)  cmd_mongodb_reset ;;
+    status) cmd_mongodb_status ;;
+    *)
+      echo "Unknown mongodb command: $cmd" >&2
+      echo "Usage: ./start-services.sh mongodb [start|stop|reset|status]" >&2
+      exit 1
+      ;;
+  esac
 }
 
 parse_args() {
@@ -74,6 +139,12 @@ parse_args() {
   export FORCE_BUILD LOCAL_ATLAS
 }
 
+if [[ "${1:-}" == "mongodb" ]]; then
+  shift
+  run_mongodb_subcommand "${1:-start}"
+  exit 0
+fi
+
 parse_args "$@"
 
 if ! command -v docker >/dev/null 2>&1; then
@@ -81,27 +152,21 @@ if ! command -v docker >/dev/null 2>&1; then
   exit 1
 fi
 
-if docker compose version >/dev/null 2>&1; then
-  DOCKER_COMPOSE=(docker compose)
-elif command -v docker-compose >/dev/null 2>&1; then
-  DOCKER_COMPOSE=(docker-compose)
-else
-  echo "Docker Compose is not available." >&2
-  exit 1
-fi
-
-COMPOSE_FILES=(-f docker-compose.yml)
+compose_detect
+compose_files
 PROFILES=()
 
 if [[ "${RAG_DEV_STACK:-}" == "1" ]]; then
-  COMPOSE_FILES+=(-f docker-compose.dev.yml)
   echo "Dev stack enabled (RAG_DEV_STACK=1) — HMR + uvicorn --reload"
 fi
 
 if [[ "$LOCAL_ATLAS" == "1" ]]; then
-  COMPOSE_FILES+=(-f docker-compose.local-atlas.yml)
-  PROFILES+=(--profile local-atlas)
+  compose_export_local_atlas_env
+  compose_local_atlas_profiles
+  PROFILES+=("${COMPOSE_PROFILES[@]}")
   echo "Local Atlas enabled (--local) — mongodb-atlas-local container, no cloud account needed"
+else
+  compose_clear_local_atlas_env
 fi
 
 ensure_env() {
@@ -124,7 +189,7 @@ ensure_env() {
   source .env
   set +a
 
-  # Local Atlas: server MONGODB_URI is set by docker-compose.local-atlas.yml overlay.
+  # Local Atlas: server MONGODB_URI is set via RAG_SERVER_MONGODB_URI env (compose_export_local_atlas_env).
   # Cloud Atlas: .env MONGODB_URI must be a real connection string.
   if [[ "$LOCAL_ATLAS" == "0" ]]; then
     if [[ -z "${MONGODB_URI:-}" ]] || [[ "$MONGODB_URI" == *"your_mongodb_atlas_uri_here"* ]]; then
@@ -189,7 +254,7 @@ print_unhealthy_server_hint() {
   if [[ "$LOCAL_ATLAS" == "1" ]]; then
     echo "Local Atlas hints:"
     echo "  docker logs rag-params-finder-mongodb-local 2>&1 | tail -20"
-    echo "  ./scripts/local-atlas.sh status"
+    echo "  ./start-services.sh mongodb status"
   else
     echo "Common Atlas fixes (TLS/SSL errors affect host and Docker alike):"
     echo "  • Network Access → allow your IP (curl https://api.ipify.org) or 0.0.0.0/0 for dev"
@@ -246,12 +311,12 @@ if [[ "$LOCAL_ATLAS" == "1" ]]; then
   echo "  MongoDB:   localhost:27017  (Atlas Local — no cloud quota)"
   echo ""
   echo "CLI (local Atlas — set URI for host commands):"
-  echo "  export MONGODB_URI=\"$LOCAL_URI\""
+  echo "  export MONGODB_URI=\"$RAG_LOCAL_MONGODB_URI_HOST\""
   echo "  rag-params-finder run --config configs/example-mongodb-local.yaml"
   echo ""
   echo "Switch back to cloud:  ./start-services.sh  (no --local flag)"
-  echo "Manage MongoDB only:   ./scripts/local-atlas.sh [start|stop|reset|status]"
-  echo "Reset all data:        docker compose -f docker-compose.yml --profile local-atlas down -v"
+  echo "Manage MongoDB only:   ./start-services.sh mongodb [start|stop|reset|status]"
+  echo "Reset all data:        docker compose --profile local-atlas down -v"
 else
   echo ""
   echo "CLI:       rag-params-finder run --config configs/example-mongodb-local.yaml"
